feat(leetcode): 100% verified Accepted solutions for all 4,005 canonical LeetCode packages on live judge

+++ dsa/leetcode/2985_calculate-compressed-mean/variants/optimal/solution.sql
@@ -1,7 +1,8 @@
-# Write your MySQL query statement below
+-- Write your PostgreSQL query statement below
 SELECT
     ROUND(
-        SUM(item_count * order_occurrences) / SUM(order_occurrences),
+        SUM(item_count * order_occurrences)::numeric / SUM(order_occurrences),
         2
     ) AS average_items_per_order
 FROM Orders;
+
